# specs/1-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RAG Chatbot API
  description: API for the embedded RAG chatbot in the Physical AI & Humanoid Robotics book
  version: 1.0.0
servers:
  - url: https://api.rag-chatbot.example.com/v1
    description: Production server
  - url: https://staging-api.rag-chatbot.example.com/v1
    description: Staging server

paths:
  /chat/start:
    post:
      summary: Start a new chat session
      description: Creates a new conversation session for the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_preferences:
                  type: object
                  description: User preferences for the session
                  properties:
                    response_style:
                      type: string
                      enum: [formal, casual, detailed, concise]
                    expertise_level:
                      type: string
                      enum: [beginner, intermediate, expert]
      responses:
        '201':
          description: New chat session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                    description: Unique identifier for the session
                  created_at:
                    type: string
                    format: date-time
                    description: When the session was created
        '400':
          $ref: '#/components/responses/BadRequest'

  /chat/{session_id}/query:
    post:
      summary: Submit a query to the chatbot
      description: Process a user query and return a response based on book content or user-provided text
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: The user's query
                context_mode:
                  type: string
                  enum: [book_content, user_text, mixed]
                  default: book_content
                  description: Whether to query book content, user text, or both
                user_text:
                  type: string
                  description: Optional user-provided text to query against
                response_preferences:
                  type: object
                  description: Preferences for the response format
                  properties:
                    include_citations:
                      type: boolean
                      default: true
                    response_length:
                      type: string
                      enum: [short, medium, long]
                      default: medium
      responses:
        '200':
          description: Query processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  response_id:
                    type: string
                    format: uuid
                  content:
                    type: string
                    description: The chatbot's response
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/SourceCitation'
                  processing_time:
                    type: number
                    description: Time taken to process the query in seconds
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /chat/{session_id}/history:
    get:
      summary: Get conversation history
      description: Retrieve the history of queries and responses for a session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Conversation history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationEntry'
        '404':
          $ref: '#/components/responses/NotFound'

  /documents/chunk:
    post:
      summary: Create document chunks
      description: Process and chunk document content for vector storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - source_id
              properties:
                content:
                  type: string
                  description: The document content to be chunked
                source_id:
                  type: string
                  format: uuid
                  description: Identifier for the source document
                source_type:
                  type: string
                  enum: [book_content, user_text]
                  description: Type of source document
                chunk_size:
                  type: integer
                  default: 512
                  description: Target size for chunks in tokens
      responses:
        '201':
          description: Document successfully chunked
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunk_ids:
                    type: array
                    items:
                      type: string
                      format: uuid
                    description: IDs of created chunks
                  count:
                    type: integer
                    description: Number of chunks created
        '400':
          $ref: '#/components/responses/BadRequest'

  /health:
    get:
      summary: Health check
      description: Check the health status of the API
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy]
                  timestamp:
                    type: string
                    format: date-time

components:
  schemas:
    SourceCitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        citation_text:
          type: string
          description: The text that was cited
        source_location:
          type: string
          description: Location in the original source
        relevance_score:
          type: number
          minimum: 0
          maximum: 1
          description: Relevance score for this citation

    ConversationEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [query, response]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceCitation'

    BadRequestError:
      type: object
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input parameters"
        code:
          type: integer
          example: 400

  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'

security:
  - gemini_api_key: []

securitySchemes:
  gemini_api_key:
    type: apiKey
    in: header
    name: X-Gemini-API-Key
    description: Gemini API key for authentication